requirements.txt
autopep8==1.5.4
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
ciscoconfparse==1.5.5
colorama==0.4.3
configparser==5.0.0
cryptography==2.9.2
dataclasses==0.7
dnspython==1.16.0
docopt==0.6.2
future==0.18.2
icdiff==1.9.1
idna==2.9
importlib-metadata==1.6.1
importlib-resources==2.0.1
Jinja2==2.11.2
jsnapy==1.3.4
junos-eznc==2.2.1
lxml==4.5.1
MarkupSafe==1.1.1
mypy-extensions==0.4.3
napalm==2.5.0
ncclient==0.6.7
netaddr==0.7.20
netmiko==3.1.1
nornir==2.4.0
nxapi-plumbing==0.5.2
paramiko==2.7.1
passlib==1.7.2
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyeapi==0.8.3
pyIOSXR==0.53
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.4
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scp==0.13.2
six==1.15.0
textfsm==1.1.0
toml==0.10.1
tqdm==4.46.1
typing-extensions==3.7.4.2
urllib3==1.25.9
zipp==3.1.0
